Important changes - all features, CSS, FS
The webpage that I have done before was not adapted to the complex way of
controlling the bar. Indeed, reading the manual thoroughly made me
understand that some modes could be programmed and that some features
(modifiers) where incompatible with certain modes.

I have separated the webpage from the ino file (it was embended as a
const char[] defore). Now the webpage is only in the .html file, also
used for testing. Therefore, the risk of differences between these two
files is removed.

I have also separated the CSS part of the webpage and included jQuery
file. Because the bar might be accessed form a device without connection
to internet, I had to embed that file. jQuery is necessary because the
forms should be dynamic, meaning that some inputs visibility is
conditionned on the value of other entries.

Use of several files required the use of SPIFSS, a filesystem for
ESPx. See https://github.com/esp8266/arduino-esp8266fs-plugin. CSS, HTML
and jQuery are located in a subfolder data, uploaded in the ESP memory
during a dedicated process (different from uploading the sketch).

I have not yet implemented the changed in the INO file that is why this
commit is not operable.