(WIP) Use jQuery, POST requests and ajax
Still under development and to be tested.

I am following the idea that any change of any control should
trigger a POST request. In order to avoid the refreshing and
clearing of the form, I return false.
Indeed, I have abandoned the idea that the server should send a
new page at every request, reflecting the changes. I prefer to
rely on CSS and javascript to manage the controls.

I am not able to test if my POST requests are properly sent.
I should also check that the ESP8266 interpret them properly.